services/diarization/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pyannote.audio>=3.3.0
torch>=2.2.0
python-multipart>=0.0.9
pydantic>=2.0.0
